Enhanced DetailScreen UI, Updated APK, and Code Refinements
Description:
In this commit, significant enhancements were made to the DetailScreen. The "More Info" section has been revamped to achieve a more presentable appearance, improving the overall user experience. Additionally, the release APK file has been updated to its latest version to ensure the inclusion of recent changes.

As part of the refinement process, the file "Kirusu.jpg" has been removed. This removal was carried out in order to facilitate thorough component preview testing.

Furthermore, this commit includes commented-out code for the preview functionality of composable within the PersonFullScreen and CharacterFullScreen sections. This decision was made to temporarily disable the preview function while ongoing development and adjustments are being made.