Fixing Search panel request link
chore: removed commented code in HomeScreenViewModel.kt and MalApiService.kt file for better readability.
fix: Fixed a bug related to input in the search box - when the user changes at least one character during the query, then there was no repeated request to the server, but the old one remains.The request to the server became more readable, because we removed unnecessary characters in the request to the server, for example, "genres=" always remained and it did not interfere with the response from the server, but it was less readable.